requirements/base.txt
boto3==1.17.6
celery-redbeat==2.0.0
celery[redis]==5.1
Django==3.1.8
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
python-dotenv==0.15.0
sentry-sdk==0.19.5
twilio==6.51.0